requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
heroku==0.1.4
idna==2.9
Markdown==3.2.1
psycopg2==2.8.5
python-dateutil==1.5
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
django-dotenv==1.4.2
factory-boy==3.1.0
Faker==4.14.0
six==1.15.0
text-unidecode==1.3